Changing pagination.js to keyword_searcher.js
Deleted file pagination.js
Created file keyword_searcher.js
In keyword_searcher.js bugfix done - click on icon, keypress "enter" agrees keyword to search the results
